+++ src/Calc/Matrix.h
@@ -23,10 +23,10 @@ private:
     /// The data stored in the matrix.
     /// </summary>
     std::vector<std::vector<double>> Data;
-    unsigned int rows;
-    unsigned int cols;
+    size_t rows;
+    size_t cols;
 
-    void Allocate(unsigned int NewRows, unsigned int NewColumns, double Value = 0) noexcept;
+    void Allocate(size_t NewRows, size_t NewColumns, double Value = 0) noexcept;
 
     using ColumnSchema = std::vector<std::pair<bool, unsigned long>>;
     [[nodiscard]] ColumnSchema GetColumnWidthSchematic() const noexcept;
@@ -35,21 +35,21 @@ private:
     Matrix();
 
 public:
-    Matrix(unsigned int Rows, unsigned int Columns) noexcept;
+    Matrix(size_t Rows, size_t Columns) noexcept;
     [[maybe_unused]] explicit Matrix(const MathVector& in);
     Matrix(const Matrix& Other) noexcept;
     Matrix(Matrix&& Other) noexcept;
 
     template<std::convertible_to<double>... args>
-    [[nodiscard]] static Matrix FromList(unsigned Rows, unsigned Columns, args... vals);
+    [[nodiscard]] static Matrix FromList(size_t Rows, size_t Columns, args... vals);
 
     Matrix& operator=(const Matrix& Other) noexcept;
     Matrix& operator=(Matrix&& Other) noexcept;
 
     friend std::ostream& operator<<(std::ostream&, const struct MatrixSingleLinePrint&);
 
-    [[nodiscard]] unsigned int Rows() const { return Data.size(); }
-    [[nodiscard]] unsigned int Columns() const { return Data.empty() ? 0 : Data[0].size(); }
+    [[nodiscard]] size_t Rows() const { return Data.size(); }
+    [[nodiscard]] size_t Columns() const { return Data.empty() ? 0 : Data[0].size(); }
     [[nodiscard]] bool IsValid() const { return rows != 0 && cols != 0; }
     [[nodiscard]] bool IsSquare() const { return IsValid() && rows == cols; }
 
@@ -66,18 +66,18 @@ public:
     void ui_dsp_fmt(std::ostream& out) const noexcept override;
 
     [[nodiscard]] static Matrix ErrorMatrix();
-    [[nodiscard]] static Matrix Identity(unsigned int Size);
-    [[nodiscard]] static Matrix Identity(unsigned int Rows, unsigned int Cols);
-    [[nodiscard]] [[maybe_unused]] static Matrix RandomMatrix(unsigned int Rows, unsigned int Columns, bool Integers);
+    [[nodiscard]] static Matrix Identity(size_t Size);
+    [[nodiscard]] static Matrix Identity(size_t Rows, size_t Cols);
+    [[nodiscard]] [[maybe_unused]] static Matrix RandomMatrix(size_t Rows, size_t Columns, bool Integers);
 
-    [[nodiscard]] const std::vector<double>& operator[](unsigned int Row) const;
-    [[nodiscard]] const double& Access(unsigned int i, unsigned int j) const;
-    [[nodiscard]] double& Access(unsigned int i, unsigned int j);
+    [[nodiscard]] const std::vector<double>& operator[](size_t Row) const;
+    [[nodiscard]] const double& Access(size_t i, size_t j) const;
+    [[nodiscard]] double& Access(size_t i, size_t j);
 
-    [[nodiscard]] Matrix Extract(unsigned int StartI, unsigned int StartJ, unsigned int RowCount, unsigned int ColumnCount);
+    [[nodiscard]] Matrix Extract(size_t StartI, size_t StartJ, size_t RowCount, size_t ColumnCount);
 
-    void RowSwap(unsigned int OrigRow, unsigned int NewRow);
-    void RowAdd(unsigned int OrigRow, double Fac, unsigned int TargetRow);
+    void RowSwap(size_t OrigRow, size_t NewRow);
+    void RowAdd(size_t OrigRow, double Fac, size_t TargetRow);
 
     [[maybe_unused]] [[nodiscard]] double Determinant() const;
     [[maybe_unused]] [[nodiscard]] Matrix Invert() const;
